requirements.txt
streamlit
pandas
numpy
scikit-learn
pypdf
python-docx
python-pptx
openpyxl
nltk
textblob
joblib
xgboost
mlflow
matplotlib
pytest
